cambio atributo

+++ boleta/src/main/java/com/veterinaria/boleta/dto/PagoDTO.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 @Setter
 public class PagoDTO {
     private Long idPago;
-    private Double monto;
+    private BigDecimal monto;
     private LocalDate fechaPago;
     private String metodoPago;
     private String estadoPago;
